Refactor model tests
