+++ tests/model/test_condscvi.py
@@ -1,0 +1,21 @@
+from scvi.data import synthetic_iid
+from scvi.model import CondSCVI
+
+
+def test_condscvi():
+    dataset = synthetic_iid(
+        n_labels=5,
+    )
+    CondSCVI.setup_anndata(
+        dataset,
+        "labels",
+    )
+    model = CondSCVI(dataset)
+    model.train(1, train_size=1)
+    model.get_latent_representation()
+    model.get_vamp_prior(dataset)
+
+    model = CondSCVI(dataset, weight_obs=True)
+    model.train(1, train_size=1)
+    model.get_latent_representation()
+    model.get_vamp_prior(dataset)
