+++ tests/model/test_models_with_minified_data.py
