+++ tests/model/test_multivi.py
@@ -1,0 +1,69 @@
+import numpy as np
+
+from scvi.data import synthetic_iid
+from scvi.model import MULTIVI
+
+
+def test_multivi():
+    data = synthetic_iid()
+    MULTIVI.setup_anndata(
+        data,
+        batch_key="batch",
+    )
+    vae = MULTIVI(
+        data,
+        n_genes=50,
+        n_regions=50,
+    )
+    vae.train(1, save_best=False)
+    vae.train(1, adversarial_mixing=False)
+    vae.train(3)
+    vae.get_elbo(indices=vae.validation_indices)
+    vae.get_accessibility_estimates()
+    vae.get_accessibility_estimates(normalize_cells=True)
+    vae.get_accessibility_estimates(normalize_regions=True)
+    vae.get_normalized_expression()
+    vae.get_library_size_factors()
+    vae.get_region_factors()
+    vae.get_reconstruction_error(indices=vae.validation_indices)
+    vae.get_latent_representation()
+    vae.differential_accessibility(groupby="labels", group1="label_1")
+    vae.differential_expression(groupby="labels", group1="label_1")
+
+    # Test with size factor
+    data = synthetic_iid()
+    data.obs["size_factor"] = np.random.randint(1, 5, size=(data.shape[0],))
+    MULTIVI.setup_anndata(data, batch_key="batch", size_factor_key="size_factor")
+    vae = MULTIVI(
+        data,
+        n_genes=50,
+        n_regions=50,
+    )
+    vae.train(3)
+
+    # Test with modality weights and penalties
+    data = synthetic_iid()
+    MULTIVI.setup_anndata(data, batch_key="batch")
+    vae = MULTIVI(data, n_genes=50, n_regions=50, modality_weights="cell")
+    vae.train(3)
+    vae = MULTIVI(data, n_genes=50, n_regions=50, modality_weights="universal")
+    vae.train(3)
+    vae = MULTIVI(data, n_genes=50, n_regions=50, modality_penalty="MMD")
+    vae.train(3)
+
+    # Test with non-zero protein data
+    data = synthetic_iid()
+    MULTIVI.setup_anndata(
+        data,
+        batch_key="batch",
+        protein_expression_obsm_key="protein_expression",
+        protein_names_uns_key="protein_names",
+    )
+    vae = MULTIVI(
+        data,
+        n_genes=50,
+        n_regions=50,
+        modality_weights="cell",
+    )
+    assert vae.n_proteins == data.obsm["protein_expression"].shape[1]
+    vae.train(3)
